+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [2.3.1] - 2024-11-08
+
+- a duration so far in the future or past that a date cannot be rendered throws RangeError when attempting to get expire or start date
+- duration of more than 255 chars throws error, cannot/should not read a string indefinately
+
 ## [2.3.0] - 2024-11-08
 
 - add toISOString, toJSON, and toString functions to ISOInterval, ISODate, and ISODuration